requirements.txt
aerich==0.5.8
aiogram==2.15
aiohttp==3.7.4.post0
aiosqlite==0.17.0
asgiref==3.4.1
async-timeout==3.0.1
asyncpg==0.24.0
asynctest==0.13.0
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
ddlparse==1.10.0
dictdiffer==0.9.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
emoji==1.6.0
fastapi==0.68.2
h11==0.12.0
idna==3.2
iniconfig==1.1.1
iso8601==0.1.16
jwt==1.3.1
loguru==0.5.3
multidict==5.2.0
nest-asyncio==1.5.1
packaging==21.0
pluggy==1.0.0
prompt-toolkit==2.0.10
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.5
pytest-asyncio==0.16.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
requests==2.26.0
rsa==4.7.2
schedule==1.1.0
six==1.16.0
starlette==0.14.2
toml==0.10.2
tortoise-orm==0.17.7
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
wcwidth==0.2.5
win32-setctime==1.0.3
yarl==1.6.3
